#131: Fix saving services and organizations without locations

+++ src/main/java/org/benetech/servicenet/service/impl/ImportServiceImpl.java
@@ -220,8 +220,7 @@ public class ImportServiceImpl implements ImportService {
             phone.setSrvc(service);
             phone.setLocation(location);
         });
-
-        return persistPhones(phones, service.getPhones());
+        return persistPhones(phones, service != null ? service.getPhones() : Collections.emptySet());
     }
 
     @Override
@@ -286,9 +285,11 @@ public class ImportServiceImpl implements ImportService {
         });
 
         Set<Language> common = new HashSet<>(langs);
-        common.retainAll(service.getLangs());
+        if (service != null) {
+            common.retainAll(service.getLangs());
+            service.getLangs().stream().filter(lang -> !common.contains(lang)).forEach(lang -> em.remove(lang));
+        }
 
-        service.getLangs().stream().filter(lang -> !common.contains(lang)).forEach(lang -> em.remove(lang));
         langs.stream().filter(lang -> !common.contains(lang)).forEach(lang -> em.persist(lang));
 
         return langs;
